in-place dct optimization

+++ src/dct_to_ppm.rs
@@ -20,9 +20,9 @@ lazy_static! {
 
 pub fn to_ppm(
     data: (
-        &Vec<SMatrix<u16, 8, 8>>,
-        &Vec<SMatrix<u16, 8, 8>>,
-        &Vec<SMatrix<u16, 8, 8>>,
+        &Vec<SMatrix<f32, 8, 8>>,
+        &Vec<SMatrix<f32, 8, 8>>,
+        &Vec<SMatrix<f32, 8, 8>>,
     ),
     height: u16,
     width: u16,
@@ -55,9 +55,9 @@ pub fn to_ppm(
             for j in 0..width as usize {
                 let x = j / 8;
                 let x_index = j % 8;
-                let r_val = (r_values[y][x][(i, x_index)] as f32 / 65535.0 * 255.0) as u16;
-                let g_val = (g_values[y][x][(i, x_index)] as f32 / 65535.0 * 255.0) as u16;
-                let b_val = (b_values[y][x][(i, x_index)] as f32 / 65535.0 * 255.0) as u16;
+                let r_val = (r_values[y][x][(i, x_index)] / 65535.0 * 255.0) as u16;
+                let g_val = (g_values[y][x][(i, x_index)] / 65535.0 * 255.0) as u16;
+                let b_val = (b_values[y][x][(i, x_index)] / 65535.0 * 255.0) as u16;
                 write!(file, "{} {} {} ", r_val, g_val, b_val)?;
             }
             writeln!(file)?;
